Move Tag declaration to new file
