+++ Swifter/Swifter.swift
@@ -52,44 +52,6 @@ public enum TwitterURL {
     
 }
 
-public enum UserTag {
-    case id(String)
-    case screenName(String)
-    
-    var key: String {
-        switch self {
-        case .id:           return "user_id"
-        case .screenName:   return "screen_name"
-        }
-    }
-    
-    var value: String {
-        switch self {
-        case .id(let id):           return id
-        case .screenName(let user): return user
-        }
-    }
-}
-
-public enum UsersTag {
-    case id([String])
-    case screenName([String])
-    
-    var key: String {
-        switch self {
-        case .id:           return "user_id"
-        case .screenName:   return "screen_name"
-        }
-    }
-    
-    var value: String {
-        switch self {
-        case .id(let id):           return id.joined(separator: ",")
-        case .screenName(let user): return user.joined(separator: ",")
-        }
-    }
-}
-
 public class Swifter {
     
     // MARK: - Types
